+++ R/compute_columns.R
@@ -452,7 +452,7 @@ add_crude_rate_old <- function(dt, parameters){
   # Sett .f = 2 dersom RATE ikke lar seg beregne og RATE.f ikke allerede er satt til max av TELLER.f/NEVNER.f
   dt[is.na(RATE) & RATE.f == 0, let(TELLER.f = 2, NEVNER.f = 2, RATE.f = 2, spv_tmp = 2L)]
   
-  if(parameters$MOVAVparameters$is_movav){
+  if(parameters$MOVAV$is_movav){
     dt[, (paste0("RATE", c(".fn1", ".fn3", ".fn9"))) := 0]
   }
 }